Create a dedicated crate for the extension
